requirements.txt
ansicon==1.89.0
blessed==1.19.1
certifi==2022.9.14
gunicorn==19.9.0
charset-normalizer==2.1.1
click==8.1.3
codefind==0.1.3
colorama==0.4.5
Flask==2.2.2
Flask-Cors==3.0.10
gcloud==0.18.3
googleapis-common-protos==1.56.4
httplib2==0.20.4
idna==3.4
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
jinxed==1.2.0
json-python-module==0.0.1
jurigged==0.5.2
jws==0.1.3
MarkupSafe==2.1.1
numpy==1.23.3
oauth2client==4.1.3
ovld==0.3.2
pickle-mixin==1.0.2
protobuf==4.21.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycryptodome==3.15.0
pyparsing==3.0.9
Pyrebase4==4.5.0
python-jwt==2.0.1
requests==2.28.1
requests-toolbelt==0.9.1
rsa==4.9
six==1.16.0
urllib3==1.26.12
watchdog==2.1.9
wcwidth==0.2.5
Werkzeug==2.2.2
zipp==3.8.1
